Change the default ListenIP/Port
In our Discord support channel, we often see people needing to make the
following two changes:

1. Change ListenIp to 0.0.0.0 because they don't want to set up an
reverse proxy. There is not a lot of harm in directly exposing the
server and the technically inclined can set it back to 127.0.0.1 anyway.

2. They generate a region file with the wrong port because there are two
of them: 22000 and 22023. Even though the form suggests 22000 by
default, many do fill in 22023. Help them by using the same port for
both, so they can't pick a wrong option.